+++ emiz/parse_parking_spots.py
@@ -1,8 +1,9 @@
 # coding=utf-8
-
 from collections import defaultdict
 
 import click
+
+# pylint: skip-file
 
 
 @click.command()